// src/app/models/Person.ts
export interface Person {
    _id: number;
    firstName: string;
    lastName: string;
    fullName: string;
    email: string;
    department: string;
    dui: string;
    nit: string;
    cellphone: string;
    telephone: string;
    birthDate: string;
    gender: string;
    lenguages: string;
    address: string;
    emergencyContact: string;
    licenceNumber: string;
    hireOn: string;
}